Multer

+++ backend/controllers/sauce.js
@@ -1,31 +1,66 @@
-const sauce = require('../models/sauce');
+const Sauce = require('../models/sauce');
+const fs = require('fs');
 
 exports.createSauce = (req, res, next) => {
-    delete req.body._id;
-    const sauce = new sauce({ 
-        ...req.body,
-    });
-    sauce.save() 
-    .then(() => res.status(201).json({ message: 'Objet enregistré !' }))
-    .catch(error => res.status(400).json({ error }));
-}
+    const sauceObject = JSON.parse(req.body.sauce);
+    delete sauceObject._id;
+    delete sauceObject._userId; 
+    const sauce = new Sauce({ 
+        ...sauceObject,
+        userId: req.auth.userId, 
+        imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+});
+    sauce.save()
+        .then(() => res.status(201).json({ message: 'Sauce added successfully !' }))
+        .catch(error => res.status(400).json({ error }));
+}; 
 exports.modifySauce = (req, res, next) => {
-        sauce.updateOne({ _id: req.params.id }, { ...req.body, _id: req.params.id })
-        .then(() => res.status(200).json({ message: 'Objet mis à jour !' }))
-        .catch(error => res.status(400).json({ error }));  
-}
+        const sauceObject = req.file ? {
+            ...JSON.parse(req.body.sauce),
+            imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+        } : { ...req.body };
+
+        delete sauceObject._userId;
+        sauce.findOne({_id: req.params.id})
+            .then((sauce) => {
+                if (sauce.userId != req.auth.userId) {
+                    res.status(401).json({ message : 'Not authorized'});
+                } else {
+                    sauce.updateOne({ _id: req.params.id}, { ...sauceObject, _id: req.params.id})
+                    .then(() => res.status(200).json({message : 'Objet modifié!'}))
+                    .catch(error => res.status(401).json({ error }));
+                }
+            })
+            .catch((error) => {
+                res.status(400).json({ error });
+            });
+};
 exports.deleteSauce = (req, res, next) => {
-    sauce.deleteOne({ _id: req.params.id })
-    .then(() => res.status(200).json({ message: 'Objet supprimé !' }))
-    .catch(error => res.status(400).json({ error }));
-}
+    sauce.findOne({ _id: req.params.id})
+    .then(sauce => {
+        if (sauce.userId != req.auth.userId) {
+            res.status(401).json({message: 'Not authorized'});
+        } else {
+            const filename = sauce.imageUrl.split('/images/')[1];
+            fs.unlink(`images/${filename}`, () => {
+                sauce.deleteOne({_id: req.params.id})
+                    .then(() => { res.status(200).json({message: 'Objet supprimé !'})})
+                    .catch(error => res.status(401).json({ error }));
+            });
+        }
+    })
+    .catch( error => {
+        res.status(500).json({ error });
+    });
+}; 
+
 exports.getOneSauce = (req, res, next) => { 
-    sauce.findOne({ _id: req.params.id })
+    Sauce.findOne({ _id: req.params.id })
     .then(sauce => res.status(200).json({ sauce }))
     .catch(error => res.status(400).json({ error }));
 }
 exports.getAllSauces = (req, res, next) => {
-    sauce.find()
+    Sauce.find()
     .then(sauces => res.status(200).json({ sauces }))
     .catch(error => res.status(400).json({ error }));
 }
